Move changes into `cancelOrder`
