+++ packages/sdk/src/actions/actions.ts
@@ -1,7 +1,6 @@
 import { acceptOffer } from './acceptOffer'
 import { buyToken } from './buyToken'
 import { cancelOrder } from './cancelOrder'
-import { cancelOrders } from './cancelOrders'
 import { listToken } from './listToken'
 import { placeBid } from './placeBid'
 
@@ -9,7 +8,6 @@ const actions = {
   acceptOffer,
   buyToken,
   cancelOrder,
-  cancelOrders,
   listToken,
   placeBid,
 }
